Moved whole component parts into its own structure
